% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eggSurvive.R
\name{eggSurvive}
\alias{eggSurvive}
\title{Place holder for egg survival model until it is read}
\usage{
eggSurvive(flow, temp, biomass, betas)
}
\description{
Place holder for egg survival model until it is read
}

