% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getEnvLogitPhi.R
\name{getEnvLogitPhi}
\alias{getEnvLogitPhi}
\title{Estimates the environmental part of survival by estimating daily survival,
obtaining the product on the real scale,
and converting back to the logit scale so it can be combined with length effects}
\usage{
getEnvLogitPhi(nSamples, nDays, sampleDays, phiBeta, nRivers = 4,
  nStages = 2)
}
\arguments{
\item{nSamples}{Scalar number of samples}

\item{nDays}{Scalar number of days in study/simulation period}

\item{sampleDays}{Vector of days on which samples were taken}

\item{phiBeta}{array of parameters from survival model}

\item{nRivers}{how many rivers?}

\item{nStages}{how many stages?}
}
\description{
Estimates the environmental part of survival by estimating daily survival,
obtaining the product on the real scale,
and converting back to the logit scale so it can be combined with length effects
}

