% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth.R
\name{getGrowthPerformance}
\alias{getGrowthPerformance}
\title{Estimates growth performance, i.e., the temperature part of the growth model}
\usage{
getGrowthPerformance(sampleHours, hourlyTemp, growthPars, nRivers = 4)
}
\arguments{
\item{sampleHours}{Breaks between performance periods}

\item{hourlyTemp}{Matrix of hourly temperatures by river}

\item{growthPars}{Named list of growth parameters from the model fit}

\item{nRivers}{how many rivers?}
}
\description{
Estimates growth performance, i.e., the temperature part of the growth model
}

