% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getMoveProb}
\alias{getMoveProb}
\title{Return the probability of moving to each river given a starting location}
\usage{
getMoveProb(river, moveProb)
}
\arguments{
\item{river}{Vector of starting locations}

\item{moveProb}{Matrix of season-specific movement probabilities}
}
\description{
Requires that movement probabilities are given as a nRivers x nRivers matrix, so this needs to be run for a single season.
}

