% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grow}
\alias{grow}
\title{Estimates actual growth given performance, other covariates, and parameters}
\usage{
grow(pars, river, startLength, flow, biomass, performance)
}
\arguments{
\item{pars}{Matrix of betas from growth model}

\item{river}{Vector of river locations for individuals}

\item{startLength}{Vector of starting lengths}

\item{flow}{Vector of mean flow for the growth period}

\item{biomass}{Vector of brook trout biomass during the growth period}

\item{performance}{Vector of growth performance from the \code{getGrowthPerformance()}}
}
\description{
Estimates actual growth given performance, other covariates, and parameters
}

