% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.R
\name{move}
\alias{move}
\title{Returns a vector of locations given a matrix of movement probabilities}
\usage{
move(moveProb)
}
\arguments{
\item{moveProb}{Matrix of movement probabilities where each row is the probability vector for an individual}
}
\description{
Returns a vector of locations given a matrix of movement probabilities
}

