% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruitLength.R
\name{recruitLength}
\alias{recruitLength}
\title{Placeholder until model for length at recruitment is available}
\usage{
recruitLength(river, flow, temp, biomass)
}
\arguments{
\item{river}{Vector of river locations for recruits}

\item{flow}{some measure of pre-recruitment flow}

\item{temp}{some measure of pre-recruitment temp}

\item{biomass}{density of competitors/dominators}
}
\description{
Placeholder until model for length at recruitment is available
}

