% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spawn.R
\name{spawn}
\alias{spawn}
\title{Returns number of eggs produced from fork lengths based on Letcher et al. 2007}
\usage{
spawn(forkLength, river)
}
\arguments{
\item{forkLength}{Vector of fork lengths to estimate fecundity for}
}
\description{
Returns number of eggs produced from fork lengths based on Letcher et al. 2007
}

