% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stageTransition.R
\name{stageTransition}
\alias{stageTransition}
\title{Returns a vector of stages making transitions for YOY to adults after summer}
\usage{
stageTransition(stage, season)
}
\arguments{
\item{stage}{Vector of starting stages}

\item{season}{Vector of starting season over which the transition happens}
}
\description{
Returns a vector of stages making transitions for YOY to adults after summer
}

