% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{survive}
\alias{survive}
\title{Gets a survival probability given the environmental contribution, start length, and parameters}
\usage{
survive(river, stage, forkLength, phiBeta5, envLogitPhi)
}
\arguments{
\item{river}{Vector of river locations of individuals}

\item{stage}{Vector of stages}

\item{forkLength}{Vector of lengths of individuals}

\item{phiBeta5}{Matrix of length betas}

\item{envLogitPhi}{Matrix of environmental contribution to surival from \code{getEnvLogitPhi()}}
}
\description{
Gets a survival probability given the environmental contribution, start length, and parameters
}

